pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sienge-python"
version = "0.1.0"
description = "Cliente Python para a API REST do Sienge — ERP de construcao civil"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "CONIN Engenharia", email = "amora@conin-ia.com.br" },
]
keywords = ["sienge", "erp", "construcao", "civil", "api", "wrapper", "python"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business",
]
dependencies = [
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/conin-engenharia/sienge-python"
Documentation = "https://github.com/conin-engenharia/sienge-python#readme"
Repository = "https://github.com/conin-engenharia/sienge-python"
Issues = "https://github.com/conin-engenharia/sienge-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["sienge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
